#13 [MOD] long 타입으로 area id 수정
