+++ domain/src/main/java/com/gdg/domain/entity/PlaceEntity.kt
@@ -1,7 +1,7 @@
 package com.gdg.domain.entity
 
 data class PlaceEntity(
-    val id: Int,
+    val id: Long,
     val name: String, //장소명
     val congestion: String, //인구혼잡도
     val min: Int,//인구 실시간 최소
