+++ feature/src/main/java/com/gdg/feature/detail/DetailViewModel.kt
@@ -20,10 +20,6 @@ class DetailViewModel @Inject constructor(
         MutableStateFlow(UiState.Empty)
     val getWeatherState: StateFlow<UiState<WeatherEntity>> get() = _getWeatherState
 
-    private val _getCongestionState: MutableStateFlow<UiState<CongestionEntity>> =
-        MutableStateFlow(UiState.Empty)
-    val getCongestionState: StateFlow<UiState<CongestionEntity>> get() = _getCongestionState
-
     fun getWeather(areaId: Long) {
         viewModelScope.launch {
             _getWeatherState.emit(UiState.Loading)
@@ -33,20 +29,6 @@ class DetailViewModel @Inject constructor(
                 },
                 onFailure = {
                     _getWeatherState.emit(UiState.Failure(it.message.toString()))
-                }
-            )
-        }
-    }
-
-    fun getCongestion(areaId: Long) {
-        viewModelScope.launch {
-            _getCongestionState.emit(UiState.Loading)
-            crowdZeroRepository.getCongestion(areaId).fold(
-                onSuccess = {
-                    _getCongestionState.emit(UiState.Success(it))
-                },
-                onFailure = {
-                    _getCongestionState.emit(UiState.Failure(it.message.toString()))
                 }
             )
         }
